Реализовать основной сервис и связать его с сервисом статистики.

+++ main-service/src/main/java/ru/practicum/ewm/category/dao/CategoryRepository.java
@@ -4,11 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import ru.practicum.ewm.category.model.Category;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
+public interface CategoryRepository extends JpaRepository<Category, Long> {
     boolean existsByName(String name);
 
-    boolean existsById(Integer id);
+    boolean existsById(Long id);
 
     @Query("select COUNT(e) FROM Event e WHERE e.category.id = ?1")
-    int findEventsCountByCategories(Integer catId);
+    int findEventsCountByCategories(Long catId);
 }
